requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.21.4
openai==1.3.7
openpyxl==3.1.2
pandas==2.0.3
pandas-stubs==2.0.1.230501
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
types-pytz==2023.3.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
